# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API For Isu Strategis Pemda",
        "title": "Isu Strategis Pemda",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "${HOST}:${PORT}",
    "basePath": "/",
    "paths": {
        "/csf": {
            "post": {
                "description": "Create Csf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Csf"
                ],
                "summary": "Create Csf",
                "parameters": [
                    {
                        "description": "Csf",
                        "name": "csf",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CsfCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/csf/:csfId": {
            "put": {
                "description": "Update Csf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Csf"
                ],
                "summary": "Update Csf",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Csf ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Csf",
                        "name": "csf",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CsfUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/csf/:idPohon": {
            "delete": {
                "description": "Delete Csf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Csf"
                ],
                "summary": "Delete Csf",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Csf ID",
                        "name": "csfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/csf/:tahun": {
            "get": {
                "description": "Find All Csf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Csf"
                ],
                "summary": "Find All Csf",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tahun",
                        "name": "tahun",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/csf/detail/:csfId": {
            "get": {
                "description": "Find Csf By Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Csf"
                ],
                "summary": "Find Csf By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Csf ID",
                        "name": "csfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/intermediate": {
            "post": {
                "description": "Create Intermediate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Intermediate"
                ],
                "summary": "Create Intermediate",
                "parameters": [
                    {
                        "description": "Intermediate",
                        "name": "intermediate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.IntermediateCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/intermediate/:id": {
            "put": {
                "description": "Update Intermediate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Intermediate"
                ],
                "summary": "Update Intermediate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Intermediate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Intermediate",
                        "name": "intermediate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.IntermediateUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/intermediate/:pohon_id": {
            "delete": {
                "description": "Delete Intermediate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Intermediate"
                ],
                "summary": "Delete Intermediate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Intermediate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/intermediate/:tahun": {
            "get": {
                "description": "Find All Intermediate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Intermediate"
                ],
                "summary": "Find All Intermediate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tahun",
                        "name": "tahun",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/intermediate/detail/:id": {
            "get": {
                "description": "Find Intermediate By Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Intermediate"
                ],
                "summary": "Find Intermediate By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Intermediate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "web.AlasanKondisiCreateRequest": {
            "type": "object",
            "properties": {
                "alasan_kondisi_strategis": {
                    "type": "string"
                },
                "data_terukur": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.DataTerukurCreateRequest"
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "web.AlasanKondisiUpdateRequest": {
            "type": "object",
            "properties": {
                "alasan_kondisi_strategis": {
                    "type": "string"
                },
                "data_terukur": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.DataTerukurUpdateRequest"
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "web.CsfCreateRequest": {
            "type": "object",
            "required": [
                "pernyataan_kondisi_strategis",
                "pohon_id",
                "tahun"
            ],
            "properties": {
                "alasan_kondisi": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.AlasanKondisiCreateRequest"
                    }
                },
                "pernyataan_kondisi_strategis": {
                    "type": "string"
                },
                "pohon_id": {
                    "type": "integer"
                },
                "tahun": {
                    "type": "string"
                }
            }
        },
        "web.CsfUpdateRequest": {
            "type": "object",
            "required": [
                "id",
                "pernyataan_kondisi_strategis",
                "pohon_id",
                "tahun"
            ],
            "properties": {
                "alasan_kondisi": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.AlasanKondisiUpdateRequest"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "pernyataan_kondisi_strategis": {
                    "type": "string"
                },
                "pohon_id": {
                    "type": "integer"
                },
                "tahun": {
                    "type": "string"
                }
            }
        },
        "web.DataTerukurCreateRequest": {
            "type": "object",
            "properties": {
                "data_terukur": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "web.DataTerukurUpdateRequest": {
            "type": "object",
            "properties": {
                "data_terukur": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "web.IntermediateCreateRequest": {
            "type": "object",
            "required": [
                "data_terukur",
                "faktor_outcome",
                "pohon_id",
                "tahun"
            ],
            "properties": {
                "data_terukur": {
                    "type": "string"
                },
                "faktor_outcome": {
                    "type": "string"
                },
                "pohon_id": {
                    "type": "integer"
                },
                "tahun": {
                    "type": "string"
                }
            }
        },
        "web.IntermediateUpdateRequest": {
            "type": "object",
            "required": [
                "data_terukur",
                "faktor_outcome",
                "pohon_id",
                "tahun"
            ],
            "properties": {
                "data_terukur": {
                    "type": "string"
                },
                "faktor_outcome": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pohon_id": {
                    "type": "integer"
                },
                "tahun": {
                    "type": "string"
                }
            }
        },
        "web.WebResponse": {
            "description": "Web Response",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}