# model/web/intermediate_create_request.go
package web

type IntermediateCreateRequest struct {
	PohonId       int    `json:"pohon_id" validate:"required"`
	Tahun         string `json:"tahun" validate:"required"`
	FaktorOutcome string `json:"faktor_outcome" validate:"required"`
	DataTerukur   string `json:"data_terukur" validate:"required"`
}